refactor: move template methods from instance to extension
refs #783
- extensions are the only things that use the template method
- cleans up the instance class a bit
