+++ lib/instance.js
@@ -1,9 +1,7 @@
 'use strict';
 const os = require('os');
-const fs = require('fs-extra');
 const path = require('path');
 const Config = require('./utils/config');
-const Promise = require('bluebird');
 
 /**
  * Class for managing the various instance-specific
@@ -227,94 +225,6 @@ class Instance {
                 port: this.config.get('server.port'),
                 process: this.process.name
             };
-        });
-    }
-
-    /**
-     * Creates a system config file (used mainly by extensions to configure external services)
-     * Creates a local copy of the config file in <instance-dir>/system/files, then links
-     * the file into the directory needed by the service
-     *
-     * @param {string} contents Contents of the template file to create
-     * @param {string} descriptor Description of the template file (used in the prompt)
-     * @param {string} file Filename
-     * @param {string} dir Directory to link the config file into
-     * @return {bool} True if the config file was created successfully, otherwise false
-     * @method template
-     * @public
-     */
-    template(contents, descriptor, file, dir) {
-        // If `--no-prompt` is passed to the CLI or the `--verbose` flag was not passed, don't show anything
-        if (!this.ui.allowPrompt || !this.ui.verbose) {
-            return this._generateTemplate(contents, descriptor, file, dir);
-        } else {
-            return this.ui.prompt({
-                type: 'expand',
-                name: 'choice',
-                message: `Would you like to view or edit the ${descriptor} file?`,
-                default: 'n',
-                choices: [
-                    {key: 'n', name: 'No, continue', value: 'continue'},
-                    {key: 'v', name: 'View the file', value: 'view'},
-                    {key: 'e', name: 'Edit the file before generation', value: 'edit'}
-                ]
-            }).then((answer) => {
-                const choice = answer.choice;
-
-                if (choice === 'continue') {
-                    return this._generateTemplate(contents, descriptor, file, dir);
-                }
-
-                if (choice === 'view') {
-                    this.ui.log(contents);
-                    return this.template(contents, descriptor, file, dir);
-                }
-
-                /* istanbul ignore else */
-                if (choice === 'edit') {
-                    return this.ui.prompt({
-                        type: 'editor',
-                        name: 'contents',
-                        message: 'Edit the generated file',
-                        default: contents
-                    }).then((answer) => {
-                        contents = answer.contents;
-                        return this._generateTemplate(contents, descriptor, file, dir);
-                    });
-                }
-            });
-        }
-    }
-
-    /**
-     * Actually handles saving the file. Used by the template method
-     *
-     * @param {string} contents Contents of file
-     * @param {string} descriptor description of file
-     * @param {string} file Filename
-     * @param {string} dir Directory to link
-     * @return {bool} True if the file was successfully created & linked
-     */
-    _generateTemplate(contents, descriptor, file, dir) {
-        const tmplDir = path.join(this.dir, 'system', 'files');
-        const tmplFile = path.join(tmplDir, file);
-
-        const promises = [
-            () => fs.ensureDir(tmplDir),
-            () => fs.writeFile(tmplFile, contents)
-        ];
-
-        // Dir is optional, if a file just needs to be created locally
-        // so we log here
-        this.ui.success(`Creating ${descriptor} file at ${tmplFile}`);
-
-        if (dir) {
-            const outputLocation = path.join(dir, file);
-            promises.push(() => this.ui.sudo(`ln -sf ${tmplFile} ${outputLocation}`));
-        }
-
-        return Promise.each(promises, (fn) => fn()).then(() => {
-            return Promise.resolve(true);
         });
     }
 }
