+++ test/unit/instance-spec.js
@@ -1,9 +1,6 @@
 'use strict';
 const expect = require('chai').expect;
 const sinon = require('sinon');
-const tmp = require('tmp');
-const path = require('path');
-const fs = require('fs-extra');
 const createConfigStub = require('../utils/config-stub');
 
 const Instance = require('../../lib/instance');
@@ -467,139 +464,6 @@ describe('Unit: Instance', function () {
                 expect(result.name).to.equal('testing');
                 expect(result.mode).to.equal('testing');
                 expect(result.url).to.equal('localhost');
-            });
-        });
-    });
-
-    describe('template', function () {
-        it('immediately calls _generateTemplate if ui.allowPrompt is false', function () {
-            const promptStub = sinon.stub().resolves();
-            const testInstance = new Instance({
-                prompt: promptStub,
-                allowPrompt: false,
-                verbose: true
-            }, {}, '');
-            const generateStub = sinon.stub(testInstance, '_generateTemplate').resolves(true);
-
-            return testInstance.template('some contents', 'a file', 'file.txt', '/some/dir').then((result) => {
-                expect(result).to.be.true;
-                expect(promptStub.called).to.be.false;
-                expect(generateStub.calledOnce).to.be.true;
-                expect(generateStub.args[0][0]).to.equal('some contents');
-            });
-        });
-
-        it('immediately calls _generateTemplate if ui.verbose is false', function () {
-            const promptStub = sinon.stub().resolves();
-            const testInstance = new Instance({
-                prompt: promptStub,
-                allowPrompt: true,
-                verbose: false
-            }, {}, '');
-            const generateStub = sinon.stub(testInstance, '_generateTemplate').resolves(true);
-
-            return testInstance.template('some contents', 'a file', 'file.txt', '/some/dir').then((result) => {
-                expect(result).to.be.true;
-                expect(promptStub.called).to.be.false;
-                expect(generateStub.calledOnce).to.be.true;
-                expect(generateStub.args[0][0]).to.equal('some contents');
-            });
-        });
-
-        it('immediately calls _generateTemplate if ui.allowPrompt and ui.verbose is false', function () {
-            const promptStub = sinon.stub().resolves();
-            const testInstance = new Instance({
-                prompt: promptStub,
-                allowPrompt: true,
-                verbose: false
-            }, {}, '');
-            const generateStub = sinon.stub(testInstance, '_generateTemplate').resolves(true);
-
-            return testInstance.template('some contents', 'a file', 'file.txt', '/some/dir').then((result) => {
-                expect(result).to.be.true;
-                expect(promptStub.called).to.be.false;
-                expect(generateStub.calledOnce).to.be.true;
-                expect(generateStub.args[0][0]).to.equal('some contents');
-            });
-        });
-
-        it('generates template if the choice is to continue (with --verbose)', function () {
-            const promptStub = sinon.stub().resolves({choice: 'continue'});
-            const testInstance = new Instance({prompt: promptStub, allowPrompt: true, verbose: true}, {} , '');
-            const generateStub = sinon.stub(testInstance, '_generateTemplate').resolves(true);
-
-            return testInstance.template('some contents', 'a file', 'file.txt', '/some/dir').then((result) => {
-                expect(result).to.be.true;
-                expect(generateStub.calledOnce).to.be.true;
-                expect(promptStub.calledOnce).to.be.true;
-                expect(generateStub.args[0]).to.deep.equal(['some contents', 'a file', 'file.txt', '/some/dir']);
-            });
-        });
-
-        it('logs and calls template method again if choice is view (with --verbose)', function () {
-            const promptStub = sinon.stub();
-            promptStub.onCall(0).resolves({choice: 'view'});
-            promptStub.onCall(1).resolves({choice: 'continue'});
-            const logStub = sinon.stub();
-            const testInstance = new Instance({log: logStub, prompt: promptStub, allowPrompt: true, verbose: true}, {}, '');
-            const generateStub = sinon.stub(testInstance, '_generateTemplate').resolves(true);
-
-            return testInstance.template('some contents', 'a file', 'file.txt', '/some/dir').then((result) => {
-                expect(result).to.be.true;
-                expect(promptStub.calledTwice).to.be.true;
-                expect(logStub.calledOnce).to.be.true;
-                expect(logStub.args[0][0]).to.equal('some contents');
-                expect(generateStub.calledOnce).to.be.true;
-                expect(generateStub.args[0]).to.deep.equal(['some contents', 'a file', 'file.txt', '/some/dir']);
-            });
-        });
-
-        it('opens editor and generates template with contents if choice is edit (with --verbose)', function () {
-            const promptStub = sinon.stub();
-            promptStub.onCall(0).resolves({choice: 'edit'});
-            promptStub.onCall(1).resolves({contents: 'some edited contents'});
-            const testInstance = new Instance({prompt: promptStub, allowPrompt: true, verbose: true}, {}, '');
-            const generateStub = sinon.stub(testInstance, '_generateTemplate').resolves(true);
-
-            return testInstance.template('some contents', 'a file', 'file.txt', '/some/dir').then((result) => {
-                expect(result).to.be.true;
-                expect(promptStub.calledTwice).to.be.true;
-                expect(generateStub.calledOnce).to.be.true;
-                expect(generateStub.args[0][0]).to.equal('some edited contents');
-            });
-        });
-    });
-
-    describe('_generateTemplate', function () {
-        it('writes out template to correct directory but doesn\'t link if no dir is passed', function () {
-            const dir = tmp.dirSync({unsafeCleanup: true}).name;
-            const successStub = sinon.stub();
-            const testInstance = new Instance({success: successStub}, {}, dir);
-
-            return testInstance._generateTemplate('some contents', 'a file', 'file.txt').then((result) => {
-                expect(result).to.be.true;
-                const fpath = path.join(dir, 'system', 'files', 'file.txt');
-                expect(fs.existsSync(fpath)).to.be.true;
-                expect(fs.readFileSync(fpath, 'utf8')).to.equal('some contents');
-                expect(successStub.calledOnce).to.be.true;
-            });
-        });
-
-        it('writes out template and links it correctly if dir is passed', function () {
-            const dir = tmp.dirSync({unsafeCleanup: true}).name;
-            const sudoStub = sinon.stub().resolves();
-            const successStub = sinon.stub();
-            const testInstance = new Instance({sudo: sudoStub, success: successStub}, {}, dir);
-
-            return testInstance._generateTemplate('some contents', 'a file', 'file.txt', '/another/dir').then((result) => {
-                expect(result).to.be.true;
-                const fpath = path.join(dir, 'system', 'files', 'file.txt');
-                expect(fs.existsSync(fpath)).to.be.true;
-                expect(fs.readFileSync(fpath, 'utf8')).to.equal('some contents');
-                expect(sudoStub.calledOnce).to.be.true;
-                expect(sudoStub.args[0][0]).to.equal(`ln -sf ${fpath} /another/dir/file.txt`);
-                expect(successStub.calledOnce).to.be.true;
-                expect(successStub.firstCall.args[0]).to.match(/^Creating a file file at/);
             });
         });
     });
